services/embed-services/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu